# Parchi-Document-Analyze/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parchi API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ImageUpload {
  /** Base64-encoded image data */
  imageData: string;
  /** MIME type of the image (image/jpeg, image/png, etc.) */
  mimeType: string;
}

export type GenericAlternativeTier = typeof GenericAlternativeTier[keyof typeof GenericAlternativeTier];


export const GenericAlternativeTier = {
  affordable: 'affordable',
  medium: 'medium',
  expensive: 'expensive',
} as const;

export interface GenericAlternative {
  brand_name: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  price_per_tablet_pkr?: number | null;
  tier: GenericAlternativeTier;
  /** @nullable */
  note?: string | null;
}

/**
 * @nullable
 */
export type EvidenceDataEvidenceStrength = typeof EvidenceDataEvidenceStrength[keyof typeof EvidenceDataEvidenceStrength] | null;


export const EvidenceDataEvidenceStrength = {
  strong: 'strong',
  common_practice: 'common_practice',
  limited: 'limited',
} as const;

export interface EvidenceData {
  /** @nullable */
  who_essential?: boolean | null;
  common_indications?: string[];
  /** @nullable */
  evidence_strength?: EvidenceDataEvidenceStrength;
  /** @nullable */
  evidence_note?: string | null;
  /** @nullable */
  evidence_note_urdu?: string | null;
  /** @nullable */
  doctor_question_english?: string | null;
  /** @nullable */
  doctor_question_urdu?: string | null;
}

export type MedicineTimingItem = typeof MedicineTimingItem[keyof typeof MedicineTimingItem];


export const MedicineTimingItem = {
  morning: 'morning',
  afternoon: 'afternoon',
  evening: 'evening',
  night: 'night',
} as const;

/**
 * @nullable
 */
export type MedicineFoodRelation = typeof MedicineFoodRelation[keyof typeof MedicineFoodRelation] | null;


export const MedicineFoodRelation = {
  before_food: 'before_food',
  after_food: 'after_food',
  with_food: 'with_food',
  anytime: 'anytime',
} as const;

export type MedicineConfidence = typeof MedicineConfidence[keyof typeof MedicineConfidence];


export const MedicineConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Medicine {
  id: string;
  medicine_name: string;
  /** @nullable */
  standard_name?: string | null;
  /** @nullable */
  active_formula?: string | null;
  /** @nullable */
  formula_urdu?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  dosage?: string | null;
  timing?: MedicineTimingItem[];
  /** @nullable */
  food_relation?: MedicineFoodRelation;
  /** @nullable */
  duration?: string | null;
  common_side_effects?: string[];
  /** @nullable */
  important_warning?: string | null;
  /** @nullable */
  explanation_urdu?: string | null;
  generic_alternatives?: GenericAlternative[];
  evidence?: EvidenceData;
  confidence: MedicineConfidence;
  user_edited?: boolean;
}

export interface PrescriptionResult {
  medicines: Medicine[];
  disclaimer: string;
}

export interface TestSummary {
  total: number;
  normal: number;
  borderline: number;
  flagged: number;
  urgent: number;
}

export type TestValueStatus = typeof TestValueStatus[keyof typeof TestValueStatus];


export const TestValueStatus = {
  normal: 'normal',
  high: 'high',
  low: 'low',
  borderline: 'borderline',
  unclear: 'unclear',
} as const;

export type TestValueUrgency = typeof TestValueUrgency[keyof typeof TestValueUrgency];


export const TestValueUrgency = {
  routine: 'routine',
  discuss_soon: 'discuss_soon',
  discuss_urgently: 'discuss_urgently',
} as const;

export interface TestValue {
  id: string;
  test_name: string;
  /** @nullable */
  test_name_urdu?: string | null;
  patient_value: string;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  normal_range?: string | null;
  status: TestValueStatus;
  urgency: TestValueUrgency;
  explanation_english: string;
  explanation_urdu: string;
  /** @nullable */
  doctor_question?: string | null;
  /** @nullable */
  doctor_question_urdu?: string | null;
}

export interface TestReportResult {
  test_values: TestValue[];
  summary: TestSummary;
  disclaimer: string;
}