# Parchi-Document-Analyze/lib/api-zod/src/generated/types/testReportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parchi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestSummary } from './testSummary';
import type { TestValue } from './testValue';

export interface TestReportResult {
  test_values: TestValue[];
  summary: TestSummary;
  disclaimer: string;
}