# Parchi-Document-Analyze/lib/api-zod/src/generated/types/testValue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Parchi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestValueStatus } from './testValueStatus';
import type { TestValueUrgency } from './testValueUrgency';

export interface TestValue {
  id: string;
  test_name: string;
  /** @nullable */
  test_name_urdu?: string | null;
  patient_value: string;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  normal_range?: string | null;
  status: TestValueStatus;
  urgency: TestValueUrgency;
  explanation_english: string;
  explanation_urdu: string;
  /** @nullable */
  doctor_question?: string | null;
  /** @nullable */
  doctor_question_urdu?: string | null;
}